merging

+++ fLibrary/activation_funcs.f90
@@ -1,25 +1,26 @@
 module activation_functions
+    use iso_c_binding
     implicit none
     
 contains
 
     FUNCTION sigmoid(x) result(output)
-        real, intent(in) :: x(:)
-        real :: output(size(x))
+        REAL (c_double), intent(in) :: x(:)
+        REAL (c_double) :: output(size(x))
         
         output = 1 / (1 + exp(-1 * x))
     END FUNCTION sigmoid
 
     FUNCTION sigmoid2d(x) result(output)
-        real, intent(in) :: x(:,:)
-        real :: output(size(x,1), size(x,2))
+        REAL (c_double), intent(in) :: x(:,:)
+        REAL (c_double) :: output(size(x,1), size(x,2))
         
         output = 1 / (1 + exp(-1 * x))
     END FUNCTION sigmoid2d
     
     FUNCTION relu(x) result(output)
-        real, intent(in) :: x(:)
-        real :: output(size(x))
+        REAL (c_double), intent(in) :: x(:)
+        REAL (c_double) :: output(size(x))
         
         where (x < 0)
             output = 0
@@ -29,8 +30,8 @@ contains
     END FUNCTION relu
 
     FUNCTION relu2d(x) result(output)
-        real, intent(in) :: x(:,:)
-        real :: output(size(x,1), size(x,2))
+        REAL (c_double), intent(in) :: x(:,:)
+        REAL (c_double) :: output(size(x,1), size(x,2))
         
         where (x < 0)
             output = 0
@@ -40,8 +41,8 @@ contains
     END FUNCTION relu2d
 
     FUNCTION relu4d(x) result(output)
-        real, intent(in) :: x(:,:,:,:)
-        real :: output(size(x,1), size(x,2), size(x,3), size(x,4))
+        REAL (c_double), intent(in) :: x(:,:,:,:)
+        REAL (c_double) :: output(size(x,1), size(x,2), size(x,3), size(x,4))
         
         where (x < 0)
             output = 0
@@ -51,14 +52,14 @@ contains
     END FUNCTION relu4d
 
     FUNCTION tanhh(x) result(output)
-        real, intent(in) :: x(:)
-        real :: output(size(x))
+        REAL (c_double), intent(in) :: x(:)
+        REAL (c_double) :: output(size(x))
         output = (exp(x)-exp(-1*x))/(exp(x)+exp(-1*x))
     END FUNCTION tanhh
 
     FUNCTION tanhh2d(x) result(output)
-        real, intent(in) :: x(:,:)
-        real :: output(size(x,1),size(x,2))
+        REAL (c_double), intent(in) :: x(:,:)
+        REAL (c_double) :: output(size(x,1),size(x,2))
         output = (exp(x)-exp(-1*x))/(exp(x)+exp(-1*x))
     END FUNCTION tanhh2d
